@@ -1,21 +1,22 @@
-import {BrowserRouter,Switch,Route} from 'react-router-dom'
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from './Home/Home';
-import AboutUs from "./AboutUs/AboutUs"
-import Counsulting from './Counsulting/Counsulting'
-import MobileAppDevelopment from './MobileAppDevelopment/MobileAppDevelopment'
-import ContactUs from './ContactUS/ContactUs'
+import AboutUs from './AboutUs/AboutUs';
+import Consulting from './Counsulting/Counsulting';
+import MobileAppDevelopment from './MobileAppDevelopment/MobileAppDevelopment';
+import ContactUs from './ContactUS/ContactUs';
+
 function App() {
   return (
     <div>
-     <BrowserRouter>
-     <Switch>
-      <Route exact path="/" component={Home}/>
-      <Route exact path="/AboutUs" component={AboutUs}/>
-      <Route exact path="/Consulting" component={Counsulting}/>
-      <Route exact path="/MobileAppDevelopment" component={MobileAppDevelopment}/>
-      <Route exact path="/ContactUs" component={ContactUs}/>
-     </Switch>
-     </BrowserRouter>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/AboutUs" element={<AboutUs />} />
+          <Route path="/Consulting" element={<Consulting />} />
+          <Route path="/MobileAppDevelopment" element={<MobileAppDevelopment />} />
+          <Route path="/ContactUs" element={<ContactUs />} />
+        </Routes>
+      </BrowserRouter>
     </div>
   );
 }
